Add bookmarked answer page

+++ backend/modules/answer/answer.controller.ts
@@ -25,6 +25,7 @@ export const getAnswers = async (req: Request, res: Response, next: any) => {
         .skip(skip)
         .limit(limit)
         .populate("author", "-password")
+        .populate("question", "title")
         .populate("likeCount")
         .populate({
           path: "likes",
@@ -52,33 +53,35 @@ export const getAnswers = async (req: Request, res: Response, next: any) => {
 export const createAnswer = async (req: Request, res: Response, next: any) => {
   try {
     const { questionId, content } = req.body
-    const question = await Question.findById(questionId)
+    const question = await Question.findById(questionId).populate("bookmarks")
 
     if (!question) {
       throw new HTTPError("No matching questions found!", 404)
     }
+
+    const questionBookmarkedBy = question.bookmarks.map(
+      (bookmark) => bookmark.userId
+    )
 
     const answer = await Answer.create({
       content: content,
       question: questionId,
       author: req.user._id,
     })
+    res.send({ success: 1, answer: answer })
 
     await answer.populate("author", "-password").execPopulate()
 
-    // TODO: change to callback
     const answerNotification = await Notification.create({
       title: `New answer for your question: ${question.title}`,
       body: `${req.user.username} answered your question!`,
-      subscribers: [question.author],
+      subscribers: [question.author, ...questionBookmarkedBy],
       link: `/questions/${questionId}`,
     })
 
-    req.io
-      .to(question.author.toString())
-      .emit("notification", answerNotification)
-
-    res.send({ success: 1, answer: answer })
+    answerNotification.subscribers.forEach((subscriber) => {
+      req.io.to(subscriber.toString()).emit("notification", answerNotification)
+    })
   } catch (err) {
     next(err)
   }
@@ -93,6 +96,11 @@ export const deleteAnswer = async (req: Request, res: Response, next: any) => {
     if (!answer) {
       throw new HTTPError("No matching answers found!", 404)
     }
+
+    if (answer.author.toString() !== req.user._id.toString()) {
+      throw new HTTPError("You cannot do this action!", 401)
+    }
+
     await Answer.findByIdAndDelete(answerId)
     res.send({ success: 1, deleted: 1 })
   } catch (err) {
